+++ database/src/main/java/cm/aptoide/pt/database/realm/Store.java
@@ -16,7 +16,7 @@ import io.realm.annotations.PrimaryKey;
  */
 
 public class Store extends RealmObject {
-	@PrimaryKey private int storeId;
+	@PrimaryKey private long storeId;
 	private String iconPath;
 	private String theme;
 	private long downloads;
@@ -24,11 +24,11 @@ public class Store extends RealmObject {
 	private String username;
 	private String passwordSha1;
 
-	public int getStoreId() {
+	public long getStoreId() {
 		return storeId;
 	}
 
-	public void setStoreId(int storeId) {
+	public void setStoreId(long storeId) {
 		this.storeId = storeId;
 	}
